// types/location_type.ts
export type LocationType = 'beach' | 'mountains' | 'lake' | 'park';
export type Region = 'vinnytsia'
    | 'volyn'
    | 'dnipropetrovsk'
    | 'donetsk'
    | 'zhytomyr'
    | 'zakarpattia'
    | 'zaporizhzhia'
    | 'ivano-frankivsk'
    | 'kyiv'
    | 'kirovohrad'
    | 'luhansk'
    | 'lviv'
    | 'mykolaiv'
    | 'odesa'
    | 'poltava'
    | 'rivne'
    | 'sumy'
    | 'ternopil'
    | 'kharkiv'
    | 'kherson'
    | 'khmelnytskyi'
    | 'cherkasy'
    | 'chernivtsi'
    | 'chernihiv'
    | 'crimea';